Run the same tests for both new style and legacy authz code paths
